Refactor common logic
- Extract out common code from ListViewDeploymentsTable and TreeViewDeploymentsTable
- Update to latest ts-jest package
